// Atspm/WebUI/src/features/charts/waitTime/types.ts
// #region license
// Copyright 2026 Utah Departement of Transportation
// for WebUI - types.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion
import {
  BaseChartData,
  BaseChartOptions,
  BasePlan,
  ChartType,
  DataPoint,
} from '@/features/charts/common/types'

export interface WaitTimeChartOptions extends BaseChartOptions {
  binSize: number
}

export interface WaitTimeChartOptionsDefaults {
  binSize: { id: number; value: string; option: string }
}

export interface WaitTimePlan extends BasePlan {
  averageWaitTime: number
  maxWaitTime: number
}

export interface RawWaitTimeData extends BaseChartData {
  approachId: number
  approachDescription: string
  phaseNumber: number
  detectionTypes: string
  plans: WaitTimePlan[]
  gapOuts: DataPoint[]
  maxOuts: DataPoint[]
  forceOffs: DataPoint[]
  unknowns: DataPoint[]
  average: DataPoint[]
  volumes: DataPoint[]
  planSplits: DataPoint[]
}

export interface RawWaitTimeResponse {
  type: ChartType.WaitTime
  data: RawWaitTimeData[]
}